+++ package.json
@@ -11,8 +11,12 @@
   "dependencies": {
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv-expand": "^8.0.3",
     "express": "4.17.1",
-    "morgan": "1.10.0"
+    "joi": "^17.6.0",
+    "mongoose": "^6.3.5",
+    "morgan": "1.10.0",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "eslint": "7.19.0",
